profile for collision test

+++ 2ndSemesterGame/2ndSemesterGame/Env.h
@@ -6,7 +6,7 @@ extern GameManager CurrentGameManager;
 
 class CollisionObject : public GameObject {
 	void Update() {
-		AddForce(Vector3f(rand() % 300 - 150, rand() % 300 - 150, 0));
+		AddForce(Vector3f((float)(rand() % 300 - 150), (float)(rand() % 300 - 150), 0));
 	}
 };
